+++ src/hupper/watchdog.py
@@ -16,10 +16,13 @@ class WatchdogFileMonitor(FileSystemEventHandler, Observer, IFileMonitor):
 
     ``callback`` is a callable that accepts a path to a changed file.
 
+    ``logger`` is an :class:`hupper.interfaces.ILogger` instance.
+
     """
-    def __init__(self, callback, **kw):
+    def __init__(self, callback, logger, **kw):
         super(WatchdogFileMonitor, self).__init__()
         self.callback = callback
+        self.logger = logger
         self.paths = set()
         self.dirpaths = set()
         self.lock = threading.Lock()
@@ -30,10 +33,10 @@ class WatchdogFileMonitor(FileSystemEventHandler, Observer, IFileMonitor):
             if dirpath not in self.dirpaths:
                 try:
                     self.schedule(self, dirpath)
-                except (OSError, IOError):  # pragma: no cover
-                    # ideally we would handle this better but if the
-                    # directory is missing watchdog raises an error
-                    pass
+                except (OSError, IOError) as ex:  # pragma: no cover
+                    # watchdog raises exceptions if folders are missing
+                    # or if the ulimit is passed
+                    self.logger.error('watchdog error: ' + str(ex))
                 else:
                     self.dirpaths.add(dirpath)