+++ tests/test_it.py
@@ -15,7 +15,7 @@ def test_myapp_reloads_when_touching_ini(testapp):
     testapp.stop()
 
     assert len(testapp.response) == 2
-    assert testapp.stdout != ''
+    assert testapp.stderr != ''
 
 
 def test_myapp_reloads_when_touching_pyfile(testapp):
@@ -27,4 +27,4 @@ def test_myapp_reloads_when_touching_pyfile(testapp):
     testapp.stop()
 
     assert len(testapp.response) == 2
-    assert testapp.stdout != ''
+    assert testapp.stderr != ''